Make profile switching more idiomatic
This commit removes the profile resolver and use core Spring Boot
features to achieve a similar, more idiomatic result.

Rather than making a profile mandatory for tests, we flip the
responsibility so that everything is provided with the default
configuration, and an `app` profile provides a specific URL for the
H2 datasource of the application. Technically, the URL in
application.properties should be completely removed to rely on the
auto-configuration.

Enabling a profile on the command line can be achieved in a more
elegant way by passing the property to the surefire plugin, reusing the
existing `test.profiles` property.

Finally, we need to make sure the app profile is enabled when the app
starts and that can be easily achieved using SpringApplicationBuilder.